notessync-mcp/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
modelcontextprotocol==0.1.0
pydantic==2.9.2
authlib==1.3.1
pyyaml==6.0.2
pytest==8.3.3
httpx==0.27.2
python-jose[cryptography]==3.3.0
python-logging-json-formatter